fix discriminators
