+++ bindings/nodejs/lib/types/block/payload/internal.ts
@@ -1,8 +1,9 @@
 // Copyright 2023 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
+export * from './payload';
 export * from './milestone';
 export * from './tagged';
 export * from './transaction';
-export * from './payload';
 export * from './treasury';
+export * from './discriminator';
