+++ bindings/nodejs/lib/types/models/milestone_options/discriminator.ts
@@ -1,0 +1,22 @@
+// Copyright 2023 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+import {
+    ProtocolParamsMilestoneOption,
+    ReceiptMilestoneOption,
+    MilestoneOptionType,
+} from './internal';
+
+export const MilestoneOptionDiscriminator = {
+    property: 'type',
+    subTypes: [
+        {
+            value: ReceiptMilestoneOption,
+            name: MilestoneOptionType.Receipt as any,
+        },
+        {
+            value: ProtocolParamsMilestoneOption,
+            name: MilestoneOptionType.ProtocolParams as any,
+        },
+    ],
+};
