+++ bindings/nodejs/lib/types/models/milestone_options/internal.ts
@@ -1,6 +1,7 @@
 // Copyright 2023 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
+export * from './milestone-options';
 export * from './protocol-params-milestone-option';
 export * from './receipt-milestone-option';
-export * from './milestone-options';
+export * from './discriminator';
